Conflicts:
	link.sh

+++ link.sh
@@ -1,12 +1,26 @@
+#!/usr/bin/zsh
+
+echo "Changing shell to zsh"
+chsh -s /usr/bin/zsh
+
+echo "Replacing dotfiles with symlinks"
 for file in zshrc gitconfig tmux.conf vimrc vim git.scmbrc fonts Xresources gtkrc-2.0 i3
 do
-  rm -fv $HOME/.$file
+  rm -f $HOME/.$file
   ln -sv $HOME/.dotfiles/$file $HOME/.$file
 done
 
-rm -f $HOME/Library/Services
-ln -s $HOME/.dotfiles/services $HOME/Library/Services
+if [ -d "$HOME/Library" ]; then
+  echo "Enabling Ctrl+W on OSX"
+  rm -f $HOME/Library/Services
+  ln -s $HOME/.dotfiles/services $HOME/Library/Services
 
-rm -fv $HOME/Library/Keybindings/DefaultKeyBinding.dict
-mkdir -pv $HOME/Library/Keybindings
-ln -sv $HOME/.dotfiles/key_bindings $HOME/Library/Keybindings/DefaultKeyBinding.dict
+  rm -fv $HOME/Library/Keybindings/DefaultKeyBinding.dict
+  mkdir -pv $HOME/Library/Keybindings
+  ln -sv $HOME/.dotfiles/key_bindings $HOME/Library/Keybindings/DefaultKeyBinding.dict
+fi
+
+# Update Xresources so that new terminal settings apply right away
+if (( $+commands[xrdb] )); then
+  xrdb $HOME/.Xresources
+fi
